a call node now empties the rax register after the parameters have been loaded. this prevents redundant move instructions from the rax value being loaded, and subsequently loaded out of, one of the parameter registers
