+++ src/backend/codegen/valuegen.cpp
@@ -1,7 +1,6 @@
 #include "valuegen.hpp"
 #include "codegen.hpp"
 
-#include "../ir_analyzer/node_metadata.hpp"
 #include "dataflow.hpp"
 
 #include <sstream>
