+++ src/backend/values.cpp
@@ -1,5 +1,1 @@
 #include "values.hpp"
-
-namespace ir {
-
-}
