+++ src/ir/output/ir_emitter.hpp
@@ -4,22 +4,7 @@
 #include <vector>
 #include <cstdint>
 
-namespace ir {
-    struct root;
-
-    namespace global {
-        struct function;
-        struct extern_function;
-        struct global_string;
-
-        struct parameter;
-        enum parameter_type : uint8_t;
-    }
-
-    namespace block {
-        struct block_instruction;
-    }
-}
+#include "../../ir/node_prototypes.hpp"
 
 namespace ir::output {
     template <typename T>
